Modified freeze_modules to accept recursive modules with pyfile
The pyfile that is specified in the module metadata wasn't allowed to be
a folder. This is a problem to us, as the gdb module is not a single
python file.

These modifications allow people to add pyfile recursive dirs from
non-Lib locations.

Also modified deep_freeze.py to accept surrogates in text it receives.
Not sure why it received surrogates though, something from pygments.