Added option for disabling module checks in configure
Changed the package check macro to check for config variable before
using pkg-config. This allows us to disable unwanted modules in
the future.

This required a regeneration of the autoconf files. Because the autoconf
files are relatively old at this point and were not created with the
same autoconf version, a lot of changes have been made.