Enabled the _hmac module
The hmac module includes the hash libraries' (md5, blake2, sha1,2,3)
objects files in its ldflags. When combining the objects for _testembed
the files were specified multiple times which caused multiple
definitions of their symbols to be inserted.

This commit removes the dependency duplication, in our case we know all
of them will be compiled so _hmac doesn't need to add them, they are
already there.